+++ conan/tools/files/__init__.py
@@ -1,5 +1,5 @@
 from conan.tools.files.files import load, save, mkdir, rmdir, ftp_download, download, get, rename, \
-    chdir, unzip, replace_in_file, collect_libs, check_md5, check_sha1, check_sha256, fix_apple_shared_install_name
+    chdir, unzip, replace_in_file, collect_libs, check_md5, check_sha1, check_sha256
 
 from conan.tools.files.patches import patch, apply_conandata_patches
 from conan.tools.files.cpp_package import CppPackage
