+++ conans/assets/templates/new_v2_autotools.py
@@ -8,7 +8,7 @@ conanfile_lib = textwrap.dedent("""
     from conan import ConanFile
     from conan.tools.gnu import AutotoolsToolchain, Autotools
     from conan.tools.layout import basic_layout
-    from conan.tools.files import fix_apple_shared_install_name
+    from conan.tools.apple import fix_apple_shared_install_name
 
 
     class {package_name}Conan(ConanFile):
